Made the changes more generic.

+++ arch/xtensa/kernel/cycles.c
@@ -1,8 +1,8 @@
 #include <linux/cycles.h>
 #include <asm/uaccess.h>
 
-asmlinkage long sys_get_cycles(unsigned *val) {
-	unsigned cycles = _get_cycles();
+asmlinkage long sys_get_cycles(unsigned long *val) {
+	unsigned long cycles = _get_cycles();
 	put_user(cycles, val);
 	return 0;
 }
